Rename api package to ext {external}/fix comments/ remove backend type
