+++ ext/ext_test.go
@@ -1,4 +1,4 @@
-package api
+package ext
 
 import (
 	"errors"
