+++ ext/types.go
@@ -1,4 +1,4 @@
-package api
+package ext
 
 import (
 	"time"
